requirements.txt
aiogram
load_dotenv
psycopg2-binary